updated for 1.6.9 greenluma, and installs in the localappdata programs folder

+++ install.ps1
@@ -1,4 +1,15 @@
+# sets where the default install location is
+$local_appdata_location = $env:LOCALAPPDATA
+$install_location = $local_appdata_location + "\Programs\GreenLuma"
+
+# Check if the folder exists and tell user to run other script first if it does not exist
+if (-Not (Test-Path -Path $folderPath)) {
+    Write-Host "Run antivirus_exclusion script before running this script"
+    return
+}
+
 # user input check to make sure before running
+Write-Host "Greenluma will be installed at $install_location"
 $check = Read-Host -Prompt "Are you sure you want to install GreenLuma? [Y/N] (Default = N)"
 $check = $check.ToLower()
 
@@ -6,50 +17,57 @@ if ($check -ne "y"){
     exit
 }
 
-# catches exceptions and cancels the script from continuing
+# catches exceptions and cancels the script from continuing if steam is not found to be installed
 try {
     $KeyPath = "HKLM:\SOFTWARE\WOW6432Node\Valve\Steam"
     $ValueName = "InstallPath"
     # Attempt to get the item property
     $property = Get-ItemProperty -Path $KeyPath -Name $ValueName -ErrorAction Stop
     $steam_path = $($property.$ValueName)
-
-    # If successful, the property exists and its value is in $property.$ValueName
-    # Write-Host "The property '$ValueName' exists. Value: $($property.$ValueName)"
 }
 catch [System.Management.Automation.ItemNotFoundException] {
-    # If ItemNotFoundException is caught, the property does not exist
-    # Write-Host "The property '$ValueName' does not exist at '$KeyPath'."
     Write-Host "Steam is not installed."
     return
 }
 catch {
-    # Catch any other unexpected errors
     Write-Host "An error occurred: $($_.Exception.Message)"
     return
 }
 
-Set-Location -Path $PSScriptRoot
+# Creates shortcut to greenluma folder on the desktop
+$targetFolderPath = "$install_location"
+$shortcutPath = "$env:USERPROFILE\Desktop\GreenLuma.lnk"
 
+$shell = New-Object -ComObject WScript.Shell
+$shortcut = $shell.CreateShortcut($shortcutPath)
+$shortcut.TargetPath = $targetFolderPath
+$shortcut.Save()
+
+# change working directory to greenluma folder in the localappdata programs folder
+Set-Location -Path $install_location
+
+# downloads required files 7zip and password protected zip with greenluma
+# hardcoded so need to manually updated everytime lol
+Invoke-WebRequest https://gluma.weeniehut.duckdns.org/7za.exe -OutFile .\7za.exe
+Invoke-WebRequest https://gluma.weeniehut.duckdns.org/GreenLuma_2025_1.6.9-Steam006.zip -OutFile .\GreenLuma_2025_1.6.6-Steam006.zip
+
+# uses 7zip to extract files to a temp directory
 Start-Process -FilePath ".\7za.exe" -ArgumentList 'x', '"GreenLuma*.zip"', '-o".\temp"', '-p"cs.rin.ru"' -Wait -NoNewWindow
 
-New-Item -Path ".\greenluma" -ItemType Directory
-Move-Item -Path ".\temp\NormalMode\DLLInjector.exe" -Destination ".\greenluma\"
-Move-Item -Path ".\temp\NormalMode\DLLInjector.ini" -Destination ".\greenluma\"
-Move-Item -Path ".\temp\NormalMode\GreenLumaSettings_2025.exe" -Destination ".\greenluma\"
-Move-Item -Path ".\temp\StealthMode\DeleteSteamAppCache.exe" -Destination ".\greenluma\"
-Move-Item -Path ".\temp\StealthMode\user32SteamFamilies.dll" -Destination ".\greenluma\gluma.dll"
+Move-Item -Path ".\temp\NormalMode\DLLInjector.exe" -Destination ".\"
+Move-Item -Path ".\temp\NormalMode\DLLInjector.ini" -Destination ".\"
+Move-Item -Path ".\temp\NormalMode\GreenLumaSettings_2025.exe" -Destination ".\"
+Move-Item -Path ".\temp\StealthMode\DeleteSteamAppCache.exe" -Destination ".\"
+Move-Item -Path ".\temp\NormalMode\GreenLuma_2025_x86.dll" -Destination ".\gluma.dll"
 
 Remove-Item -Recurse ".\temp"
 
-$path = ".\greenluma\DLLInjector.ini"
+$path = ".\DLLInjector.ini"
 $file = (Get-Content "$path")
 
-# $steam_path = Get-ItemProperty "HKLM:\SOFTWARE\WOW6432Node\Valve\Steam" | Select-Object -expand InstallPath
 $steam_exe = "`"$steam_path\steam.exe`""
 
-$current_dir = Get-Location | Select-Object -expand Path
-$dll_path = "`"$current_dir\greenluma\gluma.dll`""
+$dll_path = "`"$install_location\gluma.dll`""
 
 $file = $file.Replace("Steam.exe","$steam_exe")
 $file = $file.Replace("GreenLuma_2025_x86.dll","$dll_path")
@@ -59,5 +77,11 @@ $file = $file.Replace("WaitForProcessTermination = 1","WaitForProcessTermination
 $file = $file.Replace("EnableFakeParentProcess = 0","EnableFakeParentProcess = 1")
 $file = $file.Replace("CreateFiles = 0","CreateFiles = 1")
 $file = $file.Replace("FileToCreate_1 =","FileToCreate_1 = StealthMode.bin")
+
+$file = $file.Repalce("BootImage = GreenLuma2025_Files\BootImage.bmp","BootImage =")
+$file = $file.Repalce("BootImageWidth = 500","BootImageWidth =")
+$file = $file.Repalce("BootImageHeight = 500","BootImageHeight =")
+$file = $file.Repalce("BootImageXOffest = 240","BootImageXOffest =")
+$file = $file.Repalce("BootImageYOffest = 280","BootImageYOffest =")
 
 Set-Content -Path $path -Value $file
